+++ CHANGELOG.md
@@ -134,13 +134,13 @@ Minor Enhancements
 
 ## v0.6.4
 
-Minor Enhancements
+### Minor Enhancements
 
 1. Added `deviceWidth` and `breakpoints` readonlyvar to support mobile emulation.
 
 ## v0.6.5
 
-Decent Enhancements
+### Decent Enhancements
 
 1. Added `try-catch-finally`.
 
@@ -149,3 +149,14 @@ Decent Enhancements
 Patch
 
 1. Catching only `ZwlLangException` in `try-catch-finally`.
+
+## v0.6.7
+
+### Bug fixes:
+
+1. Allow ability to omit catch from `try-finally`
+2. The empty error (when the ex is webdriver's) problem in catch is resolved.
+
+### Enhancements
+
+1. Added `dateIsAfter`, `dateIsBefore`, `withLastDayOfMonth`, `parseNum`.